refactor DonateTests directories and file names
