+++ src/backend/DonateTests/Actions/GetScore/GetScoreTest.cs
@@ -5,18 +5,17 @@ using Core.Database.Enums;
 using Core.Database.Models.Auth;
 using Core.Database.Models.Core;
 using Core.Database.Seeders;
-using Core.Exceptions;
 using Core.Setup.Auth0;
 using Core.Setup.Enums;
 using Donate;
-using Donate.Actions.MyDonations.GetDetails;
+using Donate.Actions.MyDonations.GetScore;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using TestsBase;
 
-namespace DonateTests.Services.GetMyDonationDetailsActionTest
+namespace DonateTests.Actions.GetScore
 {
     [TestClass()]
-    public class GetMyDonationDetailsActionTest
+    public class GetScoreTest
     {
         private readonly List<Module> _usedModules = new()
         {
@@ -25,14 +24,254 @@ namespace DonateTests.Services.GetMyDonationDetailsActionTest
         };
 
         [TestMethod()]
-        public async Task GetMyDonationDetailsActionTest_GetDetails_ReturnsResponse()
+        public async Task GetScoreTest_OneDonationMultipleProducts_ReturnsScore()
         {
             using var toolkit = new TestsToolkit(_usedModules);
             var context = toolkit.Resolve<AppDbContext>();
-            var action = toolkit.Resolve<GetMyDonationDetailsAction>();
+            var action = toolkit.Resolve<GetScoreAction>();
 
-            // Arrange
-            var donationId = toolkit.Hashids.EncodeLong(1);
+            var roleId = context.Roles.First(c => c.Name == RoleName.Donor).Id;
+
+            var testingUser = new User()
+            {
+                Status = DbEntityStatus.Active,
+                FirstName = "Normal",
+                LastName = "User",
+                Email = "test@testing.com",
+                WarehouseId = null
+            };
+            context.Add(testingUser);
+
+            var testingAuth0User = new Auth0User()
+            {
+                FirstName = "Auth",
+                LastName = "Auth0User",
+                Nickname = "Auth0",
+                UpdateUser = testingUser,
+                UpdateUserId = 1,
+                UpdatedAt = DateTime.UtcNow,
+                Email = testingUser.Email,
+                Identifier = "testingIdentifier",
+                User = testingUser,
+                UserId = testingUser.Id
+            };
+            context.Add(testingAuth0User);
+
+            var testUserRole = new UserRole
+            {
+                RoleId = roleId,
+                User = testingUser,
+                UpdateUserId = 1,
+                UpdatedAt = DateTime.UtcNow,
+                Status = DbEntityStatus.Active
+            };
+            context.Add(testUserRole);
+
+            toolkit.UpdateUserInfo(new CurrentUserInfo
+            {
+                Email = testingAuth0User.Email,
+                EmailVerified = true,
+                Identifier = testingAuth0User.Identifier,
+                GivenName = testingAuth0User.FirstName,
+                FamilyName = testingAuth0User.LastName,
+                Locale = "pl",
+                Name = testingAuth0User.FirstName + testingAuth0User.LastName,
+                Nickname = testingAuth0User.Nickname,
+                UpdatedAt = DateTime.UtcNow,
+            });
+
+            var donation = new Donation
+            {
+                Id = 1,
+                CreationUserId = testingUser.Id,
+                CreationUser = testingUser,
+                CreationDate = DateTime.UtcNow,
+                WarehouseId = WarehouseSeeder.Warehouse1PL.Id,
+                Name = "DNT000001",
+                IsExpired = false,
+                IsDelivered = true,
+                IsIncludedInStock = false,
+                UpdateUserId = UserSeeder.ServiceUser.Id,
+                UpdatedAt = DateTime.UtcNow,
+                Status = DbEntityStatus.Active
+            };
+            context.Add(donation);
+
+            var donationProduct1 = new DonationProduct
+            {
+                Donation = donation,
+                ProductId = ProductSeeder.Product1Rice.Id,
+                Quantity = 10,
+                UpdateUserId = UserSeeder.ServiceUser.Id,
+                UpdatedAt = DateTime.UtcNow,
+                Status = DbEntityStatus.Active
+            };
+            context.Add(donationProduct1);
+
+            var donationProduct2 = new DonationProduct
+            {
+                Donation = donation,
+                ProductId = ProductSeeder.Product2Pasta.Id,
+                Quantity = 10,
+                UpdateUserId = UserSeeder.ServiceUser.Id,
+                UpdatedAt = DateTime.UtcNow,
+                Status = DbEntityStatus.Active
+            };
+            context.Add(donationProduct2);
+
+            var donationProduct3 = new DonationProduct
+            {
+                Donation = donation,
+                ProductId = ProductSeeder.Product3Cereals.Id,
+                Quantity = 10,
+                UpdateUserId = UserSeeder.ServiceUser.Id,
+                UpdatedAt = DateTime.UtcNow,
+                Status = DbEntityStatus.Active
+            };
+            context.Add(donationProduct3);
+
+            await context.SaveChangesAsync();
+
+            // Act
+            var result = await action.Execute();
+
+            var output = result.Value as GetScoreResponse;
+
+            // Assert
+            Assert.AreEqual(1300, output!.Score);
+        }
+
+        [TestMethod()]
+        public async Task GetScoreTest_TwoDonationsSomeProducts_ReturnsScore()
+        {
+            using var toolkit = new TestsToolkit(_usedModules);
+            var context = toolkit.Resolve<AppDbContext>();
+            var action = toolkit.Resolve<GetScoreAction>();
+
+            var roleId = context.Roles.First(c => c.Name == RoleName.Donor).Id;
+
+            var testingUser = new User()
+            {
+                Status = DbEntityStatus.Active,
+                FirstName = "Normal",
+                LastName = "User",
+                Email = "test@testing.com",
+                WarehouseId = null
+            };
+            context.Add(testingUser);
+
+            var testingAuth0User = new Auth0User()
+            {
+                FirstName = "Auth",
+                LastName = "Auth0User",
+                Nickname = "Auth0",
+                UpdateUser = testingUser,
+                UpdateUserId = 1,
+                UpdatedAt = DateTime.UtcNow,
+                Email = testingUser.Email,
+                Identifier = "testingIdentifier",
+                User = testingUser,
+                UserId = testingUser.Id
+            };
+            context.Add(testingAuth0User);
+
+            var testUserRole = new UserRole
+            {
+                RoleId = roleId,
+                User = testingUser,
+                UpdateUserId = 1,
+                UpdatedAt = DateTime.UtcNow,
+                Status = DbEntityStatus.Active
+            };
+            context.Add(testUserRole);
+
+            toolkit.UpdateUserInfo(new CurrentUserInfo
+            {
+                Email = testingAuth0User.Email,
+                EmailVerified = true,
+                Identifier = testingAuth0User.Identifier,
+                GivenName = testingAuth0User.FirstName,
+                FamilyName = testingAuth0User.LastName,
+                Locale = "pl",
+                Name = testingAuth0User.FirstName + testingAuth0User.LastName,
+                Nickname = testingAuth0User.Nickname,
+                UpdatedAt = DateTime.UtcNow,
+            });
+
+            var donation = new Donation
+            {
+                Id = 1,
+                CreationUserId = testingUser.Id,
+                CreationUser = testingUser,
+                CreationDate = DateTime.UtcNow,
+                WarehouseId = WarehouseSeeder.Warehouse1PL.Id,
+                Name = "DNT000001",
+                IsExpired = false,
+                IsDelivered = true,
+                IsIncludedInStock = false,
+                UpdateUserId = UserSeeder.ServiceUser.Id,
+                UpdatedAt = DateTime.UtcNow,
+                Status = DbEntityStatus.Active
+            };
+            context.Add(donation);
+
+            var donationProduct1 = new DonationProduct
+            {
+                Donation = donation,
+                ProductId = ProductSeeder.Product1Rice.Id,
+                Quantity = 10,
+                UpdateUserId = UserSeeder.ServiceUser.Id,
+                UpdatedAt = DateTime.UtcNow,
+                Status = DbEntityStatus.Active
+            };
+            context.Add(donationProduct1);
+
+            var donation2 = new Donation
+            {
+                Id = 2,
+                CreationUserId = testingUser.Id,
+                CreationUser = testingUser,
+                CreationDate = DateTime.UtcNow,
+                WarehouseId = WarehouseSeeder.Warehouse1PL.Id,
+                Name = "DNT000002",
+                IsExpired = false,
+                IsDelivered = true,
+                IsIncludedInStock = false,
+                UpdateUserId = UserSeeder.ServiceUser.Id,
+                UpdatedAt = DateTime.UtcNow,
+                Status = DbEntityStatus.Active
+            };
+            context.Add(donation2);
+
+            var donationProduct2 = new DonationProduct
+            {
+                Donation = donation2,
+                ProductId = ProductSeeder.Product2Pasta.Id,
+                Quantity = 40,
+                UpdateUserId = UserSeeder.ServiceUser.Id,
+                UpdatedAt = DateTime.UtcNow,
+                Status = DbEntityStatus.Active
+            };
+            context.Add(donationProduct2);
+
+            await context.SaveChangesAsync();
+
+            // Act
+            var result = await action.Execute();
+
+            var output = result.Value as GetScoreResponse;
+
+            // Assert
+            Assert.AreEqual(2500, output!.Score);
+        }
+
+        [TestMethod()]
+        public async Task GetScoreTest_OneDonationNotDelivered_Returns0()
+        {
+            using var toolkit = new TestsToolkit(_usedModules);
+            var context = toolkit.Resolve<AppDbContext>();
+            var action = toolkit.Resolve<GetScoreAction>();
+
             var roleId = context.Roles.First(c => c.Name == RoleName.Donor).Id;
 
             var testingUser = new User()
@@ -104,188 +343,22 @@ namespace DonateTests.Services.GetMyDonationDetailsActionTest
             {
                 Donation = donation,
                 ProductId = ProductSeeder.Product1Rice.Id,
-                Quantity = 1,
+                Quantity = 10,
                 UpdateUserId = UserSeeder.ServiceUser.Id,
                 UpdatedAt = DateTime.UtcNow,
                 Status = DbEntityStatus.Active
             };
             context.Add(donationProduct1);
 
-            var donationProduct2 = new DonationProduct
-            {
-                Donation = donation,
-                ProductId = ProductSeeder.Product2Pasta.Id,
-                Quantity = 10,
-                UpdateUserId = UserSeeder.ServiceUser.Id,
-                UpdatedAt = DateTime.UtcNow,
-                Status = DbEntityStatus.Active
-            };
-            context.Add(donationProduct2);
-
             await context.SaveChangesAsync();
 
             // Act
-            var executed = await action.Execute(donationId);
-            var result = executed.Value as GetMyDonationDetailsResponse;
+            var result = await action.Execute();
+
+            var output = result.Value as GetScoreResponse;
 
             // Assert
-            Assert.AreEqual("DNT000001", result!.Name);
-            Assert.AreEqual(2, result.Products.Count);
-        }
-
-        [TestMethod()]
-        public async Task GetMyDonationDetailsActionTest_NoDonationInDb_ThrowException()
-        {
-            using var toolkit = new TestsToolkit(_usedModules);
-            var context = toolkit.Resolve<AppDbContext>();
-            var action = toolkit.Resolve<GetMyDonationDetailsAction>();
-
-            // Arrange
-            var donationId = toolkit.Hashids.EncodeLong(100);
-            var roleId = context.Roles.First(c => c.Name == RoleName.Donor).Id;
-
-            var testingUser = new User()
-            {
-                Status = DbEntityStatus.Active,
-                FirstName = "Normal",
-                LastName = "User",
-                Email = "test@testing.com",
-                WarehouseId = null
-            };
-            context.Add(testingUser);
-
-            var testingAuth0User = new Auth0User()
-            {
-                FirstName = "Auth",
-                LastName = "Auth0User",
-                Nickname = "Auth0",
-                UpdateUser = testingUser,
-                UpdateUserId = 1,
-                UpdatedAt = DateTime.UtcNow,
-                Email = testingUser.Email,
-                Identifier = "testingIdentifier",
-                User = testingUser,
-                UserId = testingUser.Id
-            };
-            context.Add(testingAuth0User);
-
-            var testUserRole = new UserRole
-            {
-                RoleId = roleId,
-                User = testingUser,
-                UpdateUserId = 1,
-                UpdatedAt = DateTime.UtcNow,
-                Status = DbEntityStatus.Active
-            };
-            context.Add(testUserRole);
-
-            toolkit.UpdateUserInfo(new CurrentUserInfo
-            {
-                Email = testingAuth0User.Email,
-                EmailVerified = true,
-                Identifier = testingAuth0User.Identifier,
-                GivenName = testingAuth0User.FirstName,
-                FamilyName = testingAuth0User.LastName,
-                Locale = "pl",
-                Name = testingAuth0User.FirstName + testingAuth0User.LastName,
-                Nickname = testingAuth0User.Nickname,
-                UpdatedAt = DateTime.UtcNow,
-            });
-
-            await context.SaveChangesAsync();
-
-            // Act
-            var exception = await Assert.ThrowsExceptionAsync<ItemNotFoundException>(() => action.Execute(donationId));
-
-            // Assert
-            Assert.IsInstanceOfType(exception, typeof(ItemNotFoundException));
-            Assert.IsNotNull(exception.Message);
-        }
-
-        [TestMethod()]
-        public async Task GetMyDonationDetailsActionTest_NoDonationProductInDb_ThrowException()
-        {
-            using var toolkit = new TestsToolkit(_usedModules);
-            var context = toolkit.Resolve<AppDbContext>();
-            var action = toolkit.Resolve<GetMyDonationDetailsAction>();
-
-            // Arrange
-            var donationId = toolkit.Hashids.EncodeLong(1);
-            var roleId = context.Roles.First(c => c.Name == RoleName.Donor).Id;
-
-            var testingUser = new User()
-            {
-                Status = DbEntityStatus.Active,
-                FirstName = "Normal",
-                LastName = "User",
-                Email = "test@testing.com",
-                WarehouseId = null
-            };
-            context.Add(testingUser);
-
-            var testingAuth0User = new Auth0User()
-            {
-                FirstName = "Auth",
-                LastName = "Auth0User",
-                Nickname = "Auth0",
-                UpdateUser = testingUser,
-                UpdateUserId = 1,
-                UpdatedAt = DateTime.UtcNow,
-                Email = testingUser.Email,
-                Identifier = "testingIdentifier",
-                User = testingUser,
-                UserId = testingUser.Id
-            };
-            context.Add(testingAuth0User);
-
-            var testUserRole = new UserRole
-            {
-                RoleId = roleId,
-                User = testingUser,
-                UpdateUserId = 1,
-                UpdatedAt = DateTime.UtcNow,
-                Status = DbEntityStatus.Active
-            };
-            context.Add(testUserRole);
-
-            toolkit.UpdateUserInfo(new CurrentUserInfo
-            {
-                Email = testingAuth0User.Email,
-                EmailVerified = true,
-                Identifier = testingAuth0User.Identifier,
-                GivenName = testingAuth0User.FirstName,
-                FamilyName = testingAuth0User.LastName,
-                Locale = "pl",
-                Name = testingAuth0User.FirstName + testingAuth0User.LastName,
-                Nickname = testingAuth0User.Nickname,
-                UpdatedAt = DateTime.UtcNow,
-            });
-
-            var donation = new Donation
-            {
-                Id = 1,
-                CreationUserId = testingUser.Id,
-                CreationUser = testingUser,
-                CreationDate = DateTime.UtcNow,
-                WarehouseId = WarehouseSeeder.Warehouse1PL.Id,
-                Name = "DNT000001",
-                IsExpired = false,
-                IsDelivered = false,
-                IsIncludedInStock = false,
-                UpdateUserId = UserSeeder.ServiceUser.Id,
-                UpdatedAt = DateTime.UtcNow,
-                Status = DbEntityStatus.Active
-            };
-            context.Add(donation);
-
-            await context.SaveChangesAsync();
-
-            // Act
-            var exception = await Assert.ThrowsExceptionAsync<ItemNotFoundException>(() => action.Execute(donationId));
-
-            // Assert
-            Assert.IsInstanceOfType(exception, typeof(ItemNotFoundException));
-            Assert.IsNotNull(exception.Message);
+            Assert.AreEqual(0, output!.Score);
         }
     }
 }
