+++ src/backend/DonateTests/Actions/PostPickUpDonation/PostPickupDonationTest.cs
@@ -10,10 +10,10 @@ using Donate.Actions.PickUpDonation.PostPickUpDonation;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using TestsBase;
 
-namespace DonateTests.Actions.PostPickUpDonationActionTest
+namespace DonateTests.Actions.PostPickUpDonation
 {
     [TestClass()]
-    public class PostPickupDonationActionTest
+    public class PostPickupDonationTest
     {
         private readonly List<Module> _usedModules = new()
         {
@@ -22,7 +22,7 @@ namespace DonateTests.Actions.PostPickUpDonationActionTest
         };
 
         [TestMethod()]
-        public async Task PerformDonateActionTest_PickupDonation_ChangeIsDeliveredToTrue()
+        public async Task PerformDonateTest_PickupDonation_ChangeIsDeliveredToTrue()
         {
             using var toolkit = new TestsToolkit(_usedModules);
             var context = toolkit.Resolve<AppDbContext>();
@@ -30,11 +30,11 @@ namespace DonateTests.Actions.PostPickUpDonationActionTest
 
             // Arrange
             var donationNumber = "DNT000001";
-            var user1 = PostPickupDonationActionFixture.CreateUser("Normal", "User", WarehouseSeeder.Warehouse1PL.Id);
-            var auth0User1 = PostPickupDonationActionFixture.CreateAuth0User(user1, 1);
-            var userRole1 = PostPickupDonationActionFixture.CreateUserRole(user1, RoleSeeder.Role4WarehouseKeeper.Id);
-            var donation1 = PostPickupDonationActionFixture.CreateDonation(false, donationNumber);
-            var webPush = PostPickupDonationActionFixture.CreateUserWebPush(user1, "a", "b", "c");
+            var user1 = PostPickupDonationFixture.CreateUser("Normal", "User", WarehouseSeeder.Warehouse1PL.Id);
+            var auth0User1 = PostPickupDonationFixture.CreateAuth0User(user1, 1);
+            var userRole1 = PostPickupDonationFixture.CreateUserRole(user1, RoleSeeder.Role4WarehouseKeeper.Id);
+            var donation1 = PostPickupDonationFixture.CreateDonation(false, donationNumber);
+            var webPush = PostPickupDonationFixture.CreateUserWebPush(user1, "a", "b", "c");
 
             context.Add(user1);
             context.Add(auth0User1);
@@ -65,7 +65,7 @@ namespace DonateTests.Actions.PostPickUpDonationActionTest
         }
 
         [TestMethod()]
-        public async Task PerformDonateActionTest_PickupDonation_NoWarehouseId_ThrowsException()
+        public async Task PerformDonateTest_PickupDonation_NoWarehouseId_ThrowsException()
         {
             using var toolkit = new TestsToolkit(_usedModules);
             var context = toolkit.Resolve<AppDbContext>();
@@ -73,10 +73,10 @@ namespace DonateTests.Actions.PostPickUpDonationActionTest
 
             // Arrange
             var donationNumber = "DNT000001";
-            var user1 = PostPickupDonationActionFixture.CreateUser("Normal", "User");
-            var auth0User1 = PostPickupDonationActionFixture.CreateAuth0User(user1, 1);
-            var userRole1 = PostPickupDonationActionFixture.CreateUserRole(user1, RoleSeeder.Role4WarehouseKeeper.Id);
-            var donation1 = PostPickupDonationActionFixture.CreateDonation(false, donationNumber);
+            var user1 = PostPickupDonationFixture.CreateUser("Normal", "User");
+            var auth0User1 = PostPickupDonationFixture.CreateAuth0User(user1, 1);
+            var userRole1 = PostPickupDonationFixture.CreateUserRole(user1, RoleSeeder.Role4WarehouseKeeper.Id);
+            var donation1 = PostPickupDonationFixture.CreateDonation(false, donationNumber);
 
             context.Add(user1);
             context.Add(auth0User1);
@@ -107,7 +107,7 @@ namespace DonateTests.Actions.PostPickUpDonationActionTest
         }
 
         [TestMethod()]
-        public async Task PerformDonateActionTest_PickupDonation_NoDonationName_ThrowsException()
+        public async Task PerformDonateTest_PickupDonation_NoDonationName_ThrowsException()
         {
             using var toolkit = new TestsToolkit(_usedModules);
             var context = toolkit.Resolve<AppDbContext>();
@@ -115,10 +115,10 @@ namespace DonateTests.Actions.PostPickUpDonationActionTest
 
             // Arrange
             var notFoundDonationNumber = "HEH000009";
-            var user1 = PostPickupDonationActionFixture.CreateUser("Normal", "User", WarehouseSeeder.Warehouse1PL.Id);
-            var auth0User1 = PostPickupDonationActionFixture.CreateAuth0User(user1, 1);
-            var userRole1 = PostPickupDonationActionFixture.CreateUserRole(user1, RoleSeeder.Role4WarehouseKeeper.Id);
-            var donation1 = PostPickupDonationActionFixture.CreateDonation(false, "DNT000001");
+            var user1 = PostPickupDonationFixture.CreateUser("Normal", "User", WarehouseSeeder.Warehouse1PL.Id);
+            var auth0User1 = PostPickupDonationFixture.CreateAuth0User(user1, 1);
+            var userRole1 = PostPickupDonationFixture.CreateUserRole(user1, RoleSeeder.Role4WarehouseKeeper.Id);
+            var donation1 = PostPickupDonationFixture.CreateDonation(false, "DNT000001");
 
             context.Add(user1);
             context.Add(auth0User1);
@@ -149,7 +149,7 @@ namespace DonateTests.Actions.PostPickUpDonationActionTest
         }
 
         [TestMethod()]
-        public async Task PerformDonateActionTest_PickupDonation_IsDeliveredTrue_ThrowException()
+        public async Task PerformDonateTest_PickupDonation_IsDeliveredTrue_ThrowException()
         {
             using var toolkit = new TestsToolkit(_usedModules);
             var context = toolkit.Resolve<AppDbContext>();
@@ -157,10 +157,10 @@ namespace DonateTests.Actions.PostPickUpDonationActionTest
 
             // Arrange
             var donationNumber = "DNT000001";
-            var user1 = PostPickupDonationActionFixture.CreateUser("Normal", "User", WarehouseSeeder.Warehouse1PL.Id);
-            var auth0User1 = PostPickupDonationActionFixture.CreateAuth0User(user1, 1);
-            var userRole1 = PostPickupDonationActionFixture.CreateUserRole(user1, RoleSeeder.Role4WarehouseKeeper.Id);
-            var donation1 = PostPickupDonationActionFixture.CreateDonation(true, donationNumber);
+            var user1 = PostPickupDonationFixture.CreateUser("Normal", "User", WarehouseSeeder.Warehouse1PL.Id);
+            var auth0User1 = PostPickupDonationFixture.CreateAuth0User(user1, 1);
+            var userRole1 = PostPickupDonationFixture.CreateUserRole(user1, RoleSeeder.Role4WarehouseKeeper.Id);
+            var donation1 = PostPickupDonationFixture.CreateDonation(true, donationNumber);
 
             context.Add(user1);
             context.Add(auth0User1);
